Factor out sign-in flow
