+++ Development/OpenPassDevelopmentApp/RootViewModel.swift
@@ -111,7 +111,7 @@ class RootViewModel: ObservableObject {
                 
         Task(priority: .userInitiated) {
             do {
-                try await OpenPassManager.shared.beginSignInUXFlow()
+                try await OpenPassManager.shared.signInFlow.beginSignIn()
                 self.openPassTokens = OpenPassManager.shared.openPassTokens
                 self.error = nil
             } catch {
